Реализована проверка на простое число через Future
